feat(generator, helix): introduce optional for file name retrieval
feat(errors): clang errors are now in helix format, still missing locational token info, but that will be added soon to be able to point and underline ranges.

Added the <optional> header and implemented a method to retrieve the file name optionally based on token line. Enhances flexibility and usability for obtaining file names from tokens.